Move paging to its own file and update changelog and versions
