+++ sdk/agrifood/agrifood-farming-rest/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.0.0-beta.2 (Unreleased)
 
+### Features Added
+
+- Export pagination helper function. [#15831](https://github.com/Azure/azure-sdk-for-js/pull/15831)
 
 ## 1.0.0-beta.1 (2021-05-26)
