+++ sdk/agrifood/agrifood-farming-rest/package.json
@@ -86,8 +86,7 @@
   "dependencies": {
     "@azure/core-auth": "^1.3.0",
     "@azure-rest/core-client-paging": "1.0.0-beta.1",
-    "@azure-rest/core-client": "1.0.0-beta.4",
-    "@azure/core-paging": "^1.1.1",
+    "@azure-rest/core-client": "1.0.0-beta.5",
     "@azure/core-rest-pipeline": "^1.1.0",
     "@azure/logger": "^1.0.0",
     "tslib": "^2.2.0"
