+++ sdk/agrifood/agrifood-farming-rest/review/agrifood-farming.api.md
@@ -7,7 +7,7 @@
 import { Client } from '@azure-rest/core-client';
 import { ClientOptions } from '@azure-rest/core-client';
 import { HttpResponse } from '@azure-rest/core-client';
-import { PagedAsyncIterableIterator } from '@azure/core-paging';
+import { PagedAsyncIterableIterator } from '@azure-rest/core-client-paging';
 import { PathUncheckedResponse } from '@azure-rest/core-client';
 import { RequestParameters } from '@azure-rest/core-client';
 import { TokenCredential } from '@azure/core-auth';
