+++ sdk/agrifood/agrifood-farming-rest/src/index.ts
@@ -8,6 +8,6 @@ export * from "./models";
 export * from "./parameters";
 export * from "./responses";
 
-export { paginate, PaginateReturn, GetArrayType } from "./farmBeats";
+export { paginate, PaginateReturn, GetArrayType } from "./paging";
 
 export default FarmBeats;
