+++ sdk/core/core-client-paging-rest/package.json
@@ -63,7 +63,7 @@
   "dependencies": {
     "@azure/core-paging": "^1.1.1",
     "@azure/core-rest-pipeline": "^1.0.3",
-    "@azure-rest/core-client": "1.0.0-beta.4",
+    "@azure-rest/core-client": "1.0.0-beta.5",
     "tslib": "^2.2.0"
   },
   "devDependencies": {
