+++ sdk/core/core-client-rest/CHANGELOG.md
@@ -1,4 +1,11 @@
-# Release History
+# Release History\
+
+## 1.0.0-beta.5 (2021-06-24)
+
+### Features Added
+
+- Expose client option to set `allowInsecureConnection` to support http. [#15831](https://github.com/Azure/azure-sdk-for-js/pull/15831)
+- Add new createRestError which takes a response to create a RestError. [#15831](https://github.com/Azure/azure-sdk-for-js/pull/15831)
 
 ## 1.0.0-beta.4 (2021-05-27)
