+++ sdk/core/core-client-rest/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@azure-rest/core-client",
-  "version": "1.0.0-beta.4",
+  "version": "1.0.0-beta.5",
   "description": "Core library for interfacing with AutoRest rest level generated code",
   "sdk-type": "client",
   "main": "dist/index.js",
@@ -61,7 +61,6 @@
   "sideEffects": false,
   "prettier": "@azure/eslint-plugin-azure-sdk/prettier.json",
   "dependencies": {
-    "@azure/core-paging": "^1.1.1",
     "@azure/core-auth": "^1.3.0",
     "@azure/core-rest-pipeline": "^1.1.0",
     "tslib": "^2.2.0"
